+++ programs/programs/urgent_needs/__init__.py
@@ -1,0 +1,5 @@
+from .base import UrgentNeedFunction
+from .co import co_urgent_need_functions
+
+
+urgent_need_functions: dict[str, type[UrgentNeedFunction]] = {**co_urgent_need_functions}
