+++ programs/programs/urgent_needs/co/early_childhood_mental_health_support.py
@@ -1,0 +1,5 @@
+from .util import ChildAgeFunction
+
+
+class EarlyChildhoodMentalHealthSupport(ChildAgeFunction):
+    max_age = 5
