+++ programs/programs/urgent_needs/co/early_intervention.py
@@ -1,0 +1,5 @@
+from .util import ChildAgeFunction
+
+
+class EarlyIntervention(ChildAgeFunction):
+    max_age = 2
