+++ programs/programs/urgent_needs/co/meal_in_counties.py
@@ -1,0 +1,12 @@
+from ..base import UrgentNeedFunction
+
+
+class MealInCounties(UrgentNeedFunction):
+    dependencies = ["county"]
+    counties = ["Denver County", "Jefferson County"]
+
+    def eligible(self):
+        """
+        Household lives in Denver or Jefferson County
+        """
+        return self.screen.county in self.counties
